+++ smartcontract/src/main/java/org/ton/java/smartcontract/types/NftCollectionConfig.java
@@ -15,7 +15,7 @@ import java.math.BigInteger;
 @ToString
 public class NftCollectionConfig implements WalletConfig {
     long seqno;
-    byte mode;
+    int mode;
     long createdAt;
     Address destination;
     BigInteger amount;
