+++ smartcontract/src/main/java/org/ton/java/smartcontract/types/PaymentChannelConfig.java
@@ -15,7 +15,7 @@ import java.math.BigInteger;
 @ToString
 public class PaymentChannelConfig implements WalletConfig {
     long seqno;
-    byte mode;
+    int mode;
     long createdAt;
     Address destination;
     BigInteger amount;
